requirements.txt
aiobotocore==3.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arrow==1.4.0
astroid==4.0.4
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
binaryornot==0.6.0
blinker==1.9.0
botocore==1.42.61
cachetools==7.0.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.3
cookiecutter==2.7.1
cryptography==46.0.5
cycler==0.12.1
dask==2026.1.2
databricks-sdk==0.96.0
debugpy==1.8.20
decorator==5.2.1
dill==0.4.1
docker==7.1.0
duckdb==1.4.4
executing==2.2.1
fastapi==0.135.1
fastjsonschema==2.21.2
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
GDAL==3.11.4
geopandas==1.1.2
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.0
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.3.2
gunicorn==25.1.0
h11==0.16.0
huey==2.6.0
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
isort==8.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
locket==1.0.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
mlflow==3.10.1
mlflow-skinny==3.10.1
mlflow-tracing==3.10.1
multidict==6.7.1
narwhals==2.17.0
nbclient==0.10.4
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.3
numpy==2.4.2
nvidia-nccl-cu12==2.29.7
opentelemetry-api==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
packaging==26.0
pandas==2.3.3
parso==0.8.6
partd==1.4.2
patsy==1.0.2
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
polars==1.38.1
polars-runtime-32==1.38.1
prettytable==3.17.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pylint==4.0.5
pyogrio==0.12.1
pyparsing==3.3.2
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
ruff==0.15.6
s3fs==2026.2.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==82.0.0
shapely==2.1.2
six==1.17.0
skops==0.13.0
smmap==5.0.2
soupsieve==2.8.3
SQLAlchemy==2.0.48
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.52.1
statsmodels==0.14.6
text-unidecode==1.3
threadpoolctl==3.6.0
tomlkit==0.14.0
toolz==1.1.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uv==0.10.9
uvicorn==0.41.0
wcwidth==0.6.0
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.2
xgboost==3.2.0
yarl==1.23.0
zipp==3.23.0
